+++ react-app/src/App.js
@@ -4,8 +4,8 @@ import LoginForm from "./components/auth/LoginForm";
 import SignUpAndEditForm from "./components/auth/SignUpAndEditForm";
 import NavBar from "./components/NavBar";
 import ProtectedRoute from "./components/auth/ProtectedRoute";
-import UsersList from "./components/UsersList";
-import User from "./components/User";
+import UsersList from "./components/user/UsersList";
+import User from "./components/user/User";
 import { authenticate } from "./services/auth";
 import { library } from '@fortawesome/fontawesome-svg-core';
 // import { fab } from '@fortawesome/free-brands-svg-icons';
